Rename commands.py to cli.py
